pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nethub-runtime"
version = "0.1.0"
description = "Cross-platform runtime skeleton for HomeHub/NestHub capability resolution and environment management."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8,<3",
  "PyYAML>=6.0,<7",
  "httpx>=0.27,<1",
  "packaging>=24.1,<25",
  "psutil>=6.0,<7",
]

[tool.setuptools.packages.find]
where = ["src"]
